yolov12-inference/app/requirements.txt
boto3==1.42.3
uuid6==2025.0.1
numpy==2.3.5
pillow==12.0.0
python_multipart==0.0.20
onnxruntime==1.23.2